requirements.txt
requests>=2.0.0
beautifulsoup4>=4.0.0
pandas>=2.0.0
lxml>=4.9.0
colorama>=0.4.6
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9
schedule>=1.2.0
PyYAML>=6.0.1
psycopg2-binary>=2.9.0
